+++ R/zzz.R
@@ -12,11 +12,19 @@
   if(has_rstudio){
     has_correct_version <- rstudioapi::isAvailable("1.2")
     msg <- paste0("You are running RStudio version '", rstudioapi::getVersion(), 
-      "', waiter may not work correctly in the viewer: upgrade to the latest version",
-      "to ensure or open Shiny apps in your browser."
+      "'\n, waiter may not work correctly in the viewer: upgrade to the latest version",
+      "or open shiny applications in the browser."
     )
 
     if(!has_correct_version)
       packageStartupMessage(msg)
   }
+}
+
+.onUnload <- function(libpath){
+  options(
+    WAITER_HTML = NULL,
+    WAITER_COLOR = NULL,
+    WAITER_LOGO = NULL
+  )
 }